feat(confconf)!: improve typings
To better support type inference with libraries like typebox, add
separate generic type for the schema and the configuration itself.
This way libraries that provide a type function to get the underlying
type of a schema can be used with confconf.

BREAKING CHANGE: confconf constructor function now takes one or two type parameters